Refactor: moved tests to test_helpers.py, and fixed language for test assertion from less than to higher than for log message
