+++ examples/team_recommender/tests/example_1_text_response/test_compute_cosine_similarity.py
@@ -1,5 +1,5 @@
 import json
-import numpy as np
+
 import pytest
 from example_1_text_response.cosine_similarity import compute_alignment, compute_cosine_similarity
 from example_1_text_response.openai_embeddings import create_embedding_object
